migrate last test into wallet.rs
